pkg/features: add Streaming encoding feature gates to default set
In v1.33, the variable `defaultKubernetesFeatureGates` was removed, and all
default feature gates are now part of `defaultVersionedKubernetesFeatureGates`.

When backporting the Streaming encoding feature gates to v1.31, we missed
adding them to `defaultKubernetesFeatureGates`, which causes the
TestAllRegisteredFeaturesExpected test to fail.

This commit adds the missing feature gates to fix the test failure.

Signed-off-by: Wei Fu <weifu@microsoft.com>